+++ cms/src/hooks/useAuth.tsx
@@ -1,5 +1,6 @@
 import React, {
   createContext,
+  useCallback,
   useContext,
   useEffect,
   useMemo,
@@ -7,7 +8,8 @@ import React, {
 } from "react";
 
 import useRefreshToken from "./useRefreshToken";
-import useAxiosPrivate from "./useAxiosPrivate";
+import axios, { AxiosError, AxiosInstance } from "axios";
+import config from "../config/env.config";
 
 // TODO
 // Need to set type for createContext, useState, and user
@@ -19,6 +21,8 @@ type Login = (newToken: string) => void;
 type Logout = () => Promise<null>;
 type Authorize = () => Promise<boolean>;
 
+export type AxiosPrivateContext = AxiosInstance;
+
 export interface AuthContext {
   login: Login;
   logout: Logout;
@@ -26,30 +30,158 @@ export interface AuthContext {
   accessToken: string | null;
   setAccessToken: React.Dispatch<React.SetStateAction<string | null>>;
   authorize: Authorize;
+  axiosPrivate: AxiosInstance;
+}
+
+interface FailedRequests {
+  resolve: (value: string | PromiseLike<string>) => void;
+  reject: (reason?: any) => void;
 }
 
 const AuthContext = createContext<AuthContext | null>(null);
+const AxiosPrivateContext = createContext<AxiosPrivateContext | null>(null);
 
 const AuthProvider: React.FC<{ children: React.ReactNode }> = ({
   children,
 }) => {
-  console.group("AuthProvider running...");
-  console.groupEnd();
+  console.log("AuthProvider running...");
   const [isAuthenticated, setIsAuthenticated] = useState(false);
   const [accessToken, setAccessToken] = useState<string | null>(null);
-  // const axiosPrivate = useAxiosPrivateInit(accessToken, setAccessToken);
-  const axiosPrivate = useAxiosPrivate(accessToken, setAccessToken);
-  const refresh = useRefreshToken();
+  const [isRefreshing, setIsRefreshing] = useState(false);
+  const [failedRequests, setFailedRequests] = useState<FailedRequests[]>([]);
 
-  const login: Login = (newToken) => {
-    console.log("login from AuthProvider running...");
-    console.log("newToken:", newToken);
+  const refreshToken = useRefreshToken();
+
+  const getAccessToken = useCallback(() => {
+    return accessToken;
+  }, [accessToken]);
+
+  const updateAccessToken = useCallback((newToken: string | null) => {
     setAccessToken(newToken);
-    setIsAuthenticated(true);
-  };
+  }, []);
 
-  const logout: Logout = async () => {
-    console.log("logout from AuthProvider running...");
+  const login: Login = useCallback(
+    (newToken) => {
+      console.group("login from AuthProvider running...");
+      console.log("newToken:", newToken);
+      console.groupEnd();
+      setAccessToken(newToken);
+      setIsAuthenticated(true);
+    },
+    [setAccessToken]
+  );
+
+  const retryFailedRequests = useCallback(
+    (error: AxiosError | null = null) => {
+      failedRequests.forEach((request) => {
+        if (error) {
+          request.reject(error);
+        } else {
+          if (accessToken) {
+            request.resolve(accessToken);
+          } else {
+            // This is unlikely to happen after successful request to /auth/refresh
+            request.reject(
+              new Error("Access token is null after /auth/refresh")
+            );
+          }
+        }
+      });
+      setFailedRequests([]);
+    },
+    [failedRequests, accessToken]
+  );
+
+  const axiosPrivate = useMemo(() => {
+    const axiosInstance = axios.create({
+      baseURL: config.blogAPIBase,
+      withCredentials: true,
+      headers: {
+        "Content-Type": "application/json",
+      },
+    });
+
+    // Attaches access token to request
+    axiosInstance.interceptors.request.use(
+      (config) => {
+        console.group("useAxiosPrivate requestInterceptor");
+        console.log("config.url:", config.url);
+        console.log("requestInterceptor accessToken:", accessToken);
+        console.groupEnd();
+        const currentAccessToken = getAccessToken();
+        if (currentAccessToken) {
+          config.headers.Authorization = `Bearer ${currentAccessToken}`;
+        } else {
+          console.log("accessToken not set");
+        }
+
+        console.log("config.headers:", config.headers);
+        return config;
+      },
+      (err) => Promise.reject(err)
+    );
+
+    axiosInstance.interceptors.response.use(
+      (response) => response,
+      async (err) => {
+        const originalRequest = err.config;
+
+        if (
+          (err.request.status === 403 || err.request.status === 401) &&
+          !originalRequest._retry
+        ) {
+          originalRequest._retry = true;
+          // If there is an existing request to /auth/refresh
+          // Add latest failed request to failedRequests array
+          if (isRefreshing) {
+            return new Promise<string>((resolve, reject) => {
+              setFailedRequests((prev) => [...prev, { resolve, reject }]);
+            })
+              .then((token) => {
+                originalRequest.headers.Authorization = `Bearer ${token}`;
+                return axiosInstance(originalRequest);
+              })
+              .catch((err) => {
+                return Promise.reject(err);
+              });
+          }
+
+          // If there is no existing request to /auth/refresh
+          setIsRefreshing(true);
+
+          return new Promise(async (resolve, reject) => {
+            try {
+              const refreshResponse = await refreshToken();
+              const newAccessToken = refreshResponse.accessToken;
+              updateAccessToken(newAccessToken);
+
+              originalRequest.headers.Authorization = `Bearer ${newAccessToken}`;
+              retryFailedRequests();
+              resolve(axiosInstance(originalRequest));
+            } catch (refreshErr: any) {
+              retryFailedRequests(refreshErr);
+              setAccessToken(null);
+              reject(refreshErr);
+            } finally {
+              setIsRefreshing(false);
+            }
+          });
+        }
+      }
+    );
+
+    return axiosInstance;
+  }, [
+    getAccessToken,
+    isRefreshing,
+    setIsRefreshing,
+    setFailedRequests,
+    updateAccessToken,
+  ]);
+
+  const logout: Logout = useCallback(async () => {
+    console.group("logout from AuthProvider running...");
+    console.groupEnd();
     return axiosPrivate.post("/auth/logout").then((_resolve) => {
       return new Promise(async (resolve) => {
         setAccessToken(null);
@@ -57,19 +189,22 @@ const AuthProvider: React.FC<{ children: React.ReactNode }> = ({
         setTimeout(() => resolve(null), 0);
       });
     });
-  };
+  }, [axiosPrivate, setAccessToken, setIsAuthenticated]);
 
-  const authorize = async () => {
-    console.group("authorize running...");
+  const authorize = useCallback(async () => {
+    console.log("authorize running...");
     try {
       await axiosPrivate.get("/auth");
+      // const newAccessToken = getGlobalAccessToken();
+      const refreshResponse = await refreshToken();
+      console.log("refreshResponse:", refreshResponse.accessToken);
+      setAccessToken(refreshResponse.accessToken);
       return true;
     } catch (err) {
       console.error(err);
       if (err instanceof Error && isAuthenticated) {
-        console.log(err);
         console.log("err instanceof Error:", err instanceof Error);
-        console.groupEnd();
+        console.log("[authorize] err:", err);
         // How to throw error to Tanstack Router errorElement?
         // throwError(err);
       } else {
@@ -78,16 +213,25 @@ const AuthProvider: React.FC<{ children: React.ReactNode }> = ({
       }
       return false;
     }
-  };
+  }, [
+    axiosPrivate,
+    accessToken,
+    isAuthenticated,
+    setAccessToken,
+    setIsAuthenticated,
+  ]);
 
   useEffect(() => {
     console.log("AuthProvider mounted...");
-    // const refresh = useRefreshToken();
 
     const initAuth = async () => {
       console.log("initAuth running...");
       try {
-        const refreshResponse = await refresh();
+        const refreshResponse = await refreshToken();
+        console.log(
+          "refreshResponse.accessToken:",
+          refreshResponse.accessToken
+        );
         login(refreshResponse.accessToken);
       } catch (err) {
         console.error(err);
@@ -97,9 +241,9 @@ const AuthProvider: React.FC<{ children: React.ReactNode }> = ({
     };
 
     initAuth();
-  }, []);
+  }, [isAuthenticated, refreshToken, login, setAccessToken]);
 
-  const providerValue = useMemo(() => {
+  const AuthProviderValue = useMemo(() => {
     return {
       login,
       logout,
@@ -107,12 +251,15 @@ const AuthProvider: React.FC<{ children: React.ReactNode }> = ({
       accessToken,
       setAccessToken,
       authorize,
+      axiosPrivate,
     };
   }, [accessToken, isAuthenticated]);
 
   return (
-    <AuthContext.Provider value={providerValue}>
-      {children}
+    <AuthContext.Provider value={AuthProviderValue}>
+      <AxiosPrivateContext.Provider value={axiosPrivate}>
+        {children}
+      </AxiosPrivateContext.Provider>
     </AuthContext.Provider>
   );
 };
@@ -121,10 +268,22 @@ const useAuth = () => {
   const context = useContext(AuthContext);
 
   if (!context) {
-    throw Error("useAuth needs to be called inside AuthContext Provider.");
+    throw new Error("useAuth needs to be called inside AuthContext Provider.");
   }
 
   return context;
 };
 
-export { AuthProvider as default, useAuth };
+const useAxiosPrivate = () => {
+  const context = useContext(AxiosPrivateContext);
+
+  if (!context) {
+    throw new Error(
+      "useAxiosPrivate needs to be called inside AxiosPrivateContext Provider."
+    );
+  }
+
+  return context;
+};
+
+export { AuthProvider as default, useAuth, useAxiosPrivate };