+++ cms/src/routes/_protected.tsx
@@ -3,6 +3,7 @@ import { createFileRoute, Outlet, redirect } from "@tanstack/react-router";
 export const Route = createFileRoute("/_protected")({
   beforeLoad: async ({ context, location }) => {
     console.group("/_protected beforeLoad running...");
+    console.log(context.auth);
     if (!context.auth.isAuthenticated) {
       console.log(
         "context.auth.isAuthenticated:",